Refactor apiClient to support dynamic AuthRepository selection based on environment variables
