+++ repositories/AuthRepository.ts
@@ -1,0 +1,3 @@
+export interface AuthRepository {
+  getIdToken(): Promise<string>;
+}
