+++ repositories/mock/MockAuthRepository.ts
@@ -1,0 +1,7 @@
+import { AuthRepository } from "../AuthRepository";
+
+export class MockAuthRepository implements AuthRepository {
+  async getIdToken(): Promise<string> {
+    return "mock-token";
+  }
+}
